Policy

+++ app/Http/Controllers/Api/PostController.php
@@ -11,6 +11,7 @@ use App\Models\Post;
 use App\Services\Interfaces\PostServiceInterface;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
@@ -49,6 +50,10 @@ class PostController extends Controller
 
     public function update(UpdatePostRequest $request, int $id): JsonResponse | PostResource
     {
+        $post = $this->postService->getByIdApi($id);
+
+        Gate::authorize('update', $post);
+
         $post = $this->postService->updateApi($id, $request->validated());
 
         if (!$post) return response()->json(['message' => 'Not Found'], 404);
@@ -58,6 +63,10 @@ class PostController extends Controller
 
     public function destroy(int $id): JsonResponse
     {
+        $post = $this->postService->getByIdApi($id);
+
+        Gate::authorize('update', $post);
+
         $deleted = $this->postService->deleteApi($id);
 
         if (!$deleted) return response()->json(['message' => 'Not Found'], 404);
